+++ packages/sdks/core-js-sdk/test/httpClient.test.ts
@@ -1,5 +1,6 @@
 import { DEFAULT_BASE_API_URL } from '../src/constants';
 import createHttpClient from '../src/httpClient';
+import { getClientSessionId } from '../src/httpClient/helpers';
 import createFetchLogger from '../src/httpClient/helpers/createFetchLogger';
 import { ExtendedResponse } from '../src/httpClient/types';
 
@@ -11,6 +12,7 @@ const afterRequestHook = jest.fn();
 const descopeHeaders = {
   'x-descope-sdk-name': 'core-js',
   'x-descope-sdk-version': globalThis.BUILD_VERSION,
+  'x-descope-sdk-session-id': getClientSessionId(),
 };
 
 const httpClient = createHttpClient({
@@ -74,14 +76,36 @@ describe('httpClient', () => {
     );
   });
 
+  it('should call fetch without ? when calling "get" without params', () => {
+    httpClient.get('1/2/3', {
+      queryParams: {},
+    });
+
+    expect(mockFetch).toHaveBeenCalledWith(`http://descope.com/1/2/3`, {
+      body: undefined,
+      credentials: 'include',
+      headers: new Headers({
+        test: '123',
+        Authorization: 'Bearer 456',
+        ...descopeHeaders,
+      }),
+      method: 'GET',
+    });
+  });
+
   it('should call fetch with multiple params when calling "get"', () => {
     httpClient.get('1/2/3', {
       headers: { test2: '123' },
-      queryParams: { test2: '123', test3: '456', test4: '789' },
+      queryParams: {
+        test2: '123',
+        test3: '456',
+        test4: '789',
+        test5: `don't+forget+to@escape.urls`,
+      },
     });
 
     expect(mockFetch).toHaveBeenCalledWith(
-      'http://descope.com/1/2/3?test2=123&test3=456&test4=789',
+      `http://descope.com/1/2/3?test2=123&test3=456&test4=789&test5=don't%2Bforget%2Bto%40escape.urls`,
       {
         body: undefined,
         credentials: 'include',
@@ -138,8 +162,8 @@ describe('httpClient', () => {
           test2: '123',
           test: '123',
           Authorization: 'Bearer 456',
+          ...descopeHeaders,
           'x-descope-sdk-name': 'lulu',
-          'x-descope-sdk-version': globalThis.BUILD_VERSION,
         }),
         method: 'GET',
       },
@@ -182,6 +206,25 @@ describe('httpClient', () => {
       baseConfig: { baseHeaders: { test: '123' } },
       cookiePolicy: null,
     });
+
+    httpClient.get('1/2/3/4', {
+      headers: { test2: '123' },
+      queryParams: { test2: '123' },
+    });
+
+    expect(mockFetch).toHaveBeenCalledWith(
+      'http://descope.com/1/2/3/4?test2=123',
+      {
+        body: undefined,
+        headers: new Headers({
+          test2: '123',
+          test: '123',
+          Authorization: 'Bearer 456',
+          ...descopeHeaders,
+        }),
+        method: 'GET',
+      },
+    );
 
     httpClient.get('1/2/3', {
       headers: { test2: '123' },