+++ packages/sdks/react-sdk/src/types.ts
@@ -11,6 +11,7 @@ import RoleManagementWidget from '@descope/role-management-widget';
 import AccessKeyManagementWidget from '@descope/access-key-management-widget';
 import AuditManagementWidget from '@descope/audit-management-widget';
 import UserProfileWidget from '@descope/user-profile-widget';
+import ApplicationsPortalWidget from '@descope/applications-portal-widget';
 import createSdk from './sdk';
 
 declare global {
@@ -22,6 +23,7 @@ declare global {
       ['descope-access-key-management-widget']: AccessKeyManagementCustomElement;
       ['descope-audit-management-widget']: AuditManagementCustomElement;
       ['descope-user-profile-widget']: UserProfileCustomElement;
+      ['descope-applications-portal-widget']: ApplicationsPortalCustomElement;
     }
   }
 }
@@ -77,6 +79,10 @@ export type UserProfileCustomElement = CustomElement<
   typeof UserProfileWidget & UserProfileProps
 >;
 
+export type ApplicationsPortalCustomElement = CustomElement<
+  typeof ApplicationsPortalWidget & ApplicationsPortalProps
+>;
+
 export interface IContext {
   fetchUser: () => void;
   user: User;
@@ -130,6 +136,10 @@ export type AccessKeyManagementProps = WidgetProps;
 export type AuditManagementProps = WidgetProps;
 
 export type UserProfileProps = Omit<WidgetProps, 'tenant'> & {
+  onLogout?: (e: CustomEvent) => void;
+};
+
+export type ApplicationsPortalProps = Omit<WidgetProps, 'tenant'> & {
   onLogout?: (e: CustomEvent) => void;
 };