+++ packages/widgets/applications-portal-widget/src/lib/widget/mixins/initMixin/initMixin.ts
@@ -1,0 +1,17 @@
+import { compose, createSingletonMixin } from '@descope/sdk-helpers';
+import { debuggerMixin, themeMixin } from '@descope/sdk-mixins';
+import { initAppsListMixin } from './initComponentsMixins/initAppsListMixin';
+
+export const initMixin = createSingletonMixin(
+  <T extends CustomElementConstructor>(superclass: T) =>
+    // @ts-ignore
+    class InitMixinClass extends compose(
+      debuggerMixin,
+      themeMixin,
+      initAppsListMixin,
+    )(superclass) {
+      async init() {
+        await super.init?.();
+      }
+    },
+);