+++ packages/sdks/angular-sdk/CHANGELOG.md
@@ -2,6 +2,75 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.5.25](https://github.com/descope/descope-js/compare/angular-sdk-0.5.24...angular-sdk-0.5.25) (2024-09-29)
+
+### Dependency Updates
+
+* `user-profile-widget` updated to version `0.0.107`
+* `web-component` updated to version `3.25.3`
+## [0.5.24](https://github.com/descope/descope-js/compare/angular-sdk-0.5.23...angular-sdk-0.5.24) (2024-09-29)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.1.124`
+* `audit-management-widget` updated to version `0.1.88`
+* `role-management-widget` updated to version `0.1.122`
+* `user-management-widget` updated to version `0.4.125`
+* `user-profile-widget` updated to version `0.0.106`
+* `web-component` updated to version `3.25.2`
+* `web-js-sdk` updated to version `1.17.0`
+* `core-js-sdk` updated to version `2.26.0`
+## [0.5.23](https://github.com/descope/descope-js/compare/angular-sdk-0.5.22...angular-sdk-0.5.23) (2024-09-19)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.1.123`
+* `audit-management-widget` updated to version `0.1.87`
+* `role-management-widget` updated to version `0.1.121`
+* `user-management-widget` updated to version `0.4.124`
+* `user-profile-widget` updated to version `0.0.105`
+* `web-component` updated to version `3.25.1`
+## [0.5.22](https://github.com/descope/descope-js/compare/angular-sdk-0.5.21...angular-sdk-0.5.22) (2024-09-17)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.1.122`
+* `audit-management-widget` updated to version `0.1.86`
+* `role-management-widget` updated to version `0.1.120`
+* `user-management-widget` updated to version `0.4.123`
+* `user-profile-widget` updated to version `0.0.104`
+* `web-component` updated to version `3.25.0`
+* `web-js-sdk` updated to version `1.16.6`
+* `core-js-sdk` updated to version `2.25.1`
+## [0.5.21](https://github.com/descope/descope-js/compare/angular-sdk-0.5.20...angular-sdk-0.5.21) (2024-09-11)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.1.121`
+* `audit-management-widget` updated to version `0.1.85`
+* `role-management-widget` updated to version `0.1.119`
+* `user-management-widget` updated to version `0.4.122`
+* `user-profile-widget` updated to version `0.0.103`
+* `web-component` updated to version `3.24.2`
+* `web-js-sdk` updated to version `1.16.5`
+* `core-js-sdk` updated to version `2.25.0`
+## [0.5.20](https://github.com/descope/descope-js/compare/angular-sdk-0.5.19...angular-sdk-0.5.20) (2024-09-05)
+
+### Dependency Updates
+
+* `audit-management-widget` updated to version `0.1.84`
+## [0.5.19](https://github.com/descope/descope-js/compare/angular-sdk-0.5.18...angular-sdk-0.5.19) (2024-09-03)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.1.120`
+* `audit-management-widget` updated to version `0.1.83`
+* `role-management-widget` updated to version `0.1.118`
+* `user-management-widget` updated to version `0.4.121`
+* `user-profile-widget` updated to version `0.0.102`
+* `web-component` updated to version `3.24.1`
+* `web-js-sdk` updated to version `1.16.4`
+* `core-js-sdk` updated to version `2.24.4`
 ## [0.5.18](https://github.com/descope/descope-js/compare/angular-sdk-0.5.17...angular-sdk-0.5.18) (2024-09-02)
 
 ### Dependency Updates